Refactor layout helpers into new module
