+++ include/semantic_global.h
@@ -14,9 +14,6 @@
 #include "ir_core.h"
 #include "symtable.h"
 
-size_t layout_union_members(union_member_t *members, size_t count);
-size_t layout_struct_members(struct_member_t *members, size_t count);
-
 /* Current maximum alignment for struct packing */
 extern size_t semantic_pack_alignment;
 void semantic_set_pack(size_t align);
